Break up access_test.py into multiple files
It's getting too big.

Also fix a silly bug that came up: we were testing a team package in a team-only setup.
