+++ registry/tests/access_test.py
@@ -4,15 +4,13 @@
 Access tests
 """
 
-import hashlib
 import json
 import time
 from unittest.mock import patch
-import urllib
 
 import requests
 
-from quilt_server.const import PaymentPlan, PUBLIC, TEAM
+from quilt_server.const import PaymentPlan, PUBLIC
 from quilt_server.core import encode_node, hash_contents, GroupNode, RootNode, FileNode
 
 from .utils import mock_customer, QuiltTestCase
@@ -477,230 +475,6 @@ class AccessTestCase(QuiltTestCase):
         assert resp.status_code == requests.codes.forbidden
 
     @patch('quilt_server.views.ALLOW_ANONYMOUS_ACCESS', True)
-    def testProfile(self):
-        """
-        List all accessible packages.
-        """
-        public_pkg = "publicpkg"
-        self.put_package(self.user, public_pkg, RootNode(children=dict()), is_public=True)
-
-        # The user can see own packages.
-        resp = self.app.get(
-            '/api/profile',
-            headers={
-                'Authorization': self.user
-            }
-        )
-
-        assert resp.status_code == requests.codes.ok
-        data = json.loads(resp.data.decode('utf8'))['packages']
-
-        assert data['own'] == [
-            dict(owner=self.user, name=self.pkg, is_public=False, is_team=False),
-            dict(owner=self.user, name=public_pkg, is_public=True, is_team=False),
-        ]
-        assert data['shared'] == []
-
-
-        # Other users can't see anything.
-        sharewith = "anotheruser"
-
-        resp = self.app.get(
-            '/api/profile',
-            headers={
-                'Authorization': sharewith
-            }
-        )
-
-        assert resp.status_code == requests.codes.ok
-        data = json.loads(resp.data.decode('utf8'))['packages']
-
-        assert data['own'] == []
-        assert data['shared'] == []
-
-
-        # Users can see shared packages.
-        resp = self._share_package(self.user, self.pkg, sharewith)
-        assert resp.status_code == requests.codes.ok
-
-        resp = self.app.get(
-            '/api/profile',
-            headers={
-                'Authorization': sharewith
-            }
-        )
-
-        assert resp.status_code == requests.codes.ok
-        data = json.loads(resp.data.decode('utf8'))['packages']
-
-        assert data['own'] == []
-        assert data['shared'] == [dict(owner=self.user, name=self.pkg, is_public=False, is_team=False)]
-
-
-        # Packages that are both public and shared show up under "shared".
-        resp = self._share_package(self.user, self.pkg, PUBLIC)
-        assert resp.status_code == requests.codes.ok
-
-        resp = self.app.get(
-            '/api/profile',
-            headers={
-                'Authorization': sharewith
-            }
-        )
-
-        assert resp.status_code == requests.codes.ok
-        data = json.loads(resp.data.decode('utf8'))['packages']
-
-        assert data['own'] == []
-        assert data['shared'] == [dict(owner=self.user, name=self.pkg, is_public=True, is_team=False)]
-
-    @patch('quilt_server.views.ALLOW_ANONYMOUS_ACCESS', False)
-    @patch('quilt_server.views.ALLOW_TEAM_ACCESS', True)
-    def testTeamProfile(self):
-        """
-        Test the profile endpoint but with teams and no public access.
-        """
-        public_pkg = "publicpkg"
-        self.put_package(self.user, public_pkg, RootNode(children=dict()), is_team=True)
-
-        # The user can see own packages.
-        resp = self.app.get(
-            '/api/profile',
-            headers={
-                'Authorization': self.user
-            }
-        )
-
-        assert resp.status_code == requests.codes.ok
-        data = json.loads(resp.data.decode('utf8'))['packages']
-
-        assert data['own'] == [
-            dict(owner=self.user, name=self.pkg, is_public=False, is_team=False),
-            dict(owner=self.user, name=public_pkg, is_public=False, is_team=True),
-        ]
-        assert data['shared'] == []
-
-
-        # Other users can't see anything.
-        sharewith = "anotheruser"
-
-        resp = self.app.get(
-            '/api/profile',
-            headers={
-                'Authorization': sharewith
-            }
-        )
-
-        assert resp.status_code == requests.codes.ok
-        data = json.loads(resp.data.decode('utf8'))['packages']
-
-        assert data['own'] == []
-        assert data['shared'] == []
-
-        # Users can see shared packages.
-        resp = self._share_package(self.user, self.pkg, sharewith)
-        assert resp.status_code == requests.codes.ok
-
-        resp = self.app.get(
-            '/api/profile',
-            headers={
-                'Authorization': sharewith
-            }
-        )
-
-        assert resp.status_code == requests.codes.ok
-        data = json.loads(resp.data.decode('utf8'))['packages']
-
-        assert data['own'] == []
-        assert data['shared'] == [dict(owner=self.user, name=self.pkg, is_public=False, is_team=False)]
-
-
-        # Packages that are both team and shared show up under "shared".
-        resp = self._share_package(self.user, self.pkg, TEAM)
-        assert resp.status_code == requests.codes.ok
-
-        resp = self.app.get(
-            '/api/profile',
-            headers={
-                'Authorization': sharewith
-            }
-        )
-
-        assert resp.status_code == requests.codes.ok
-        data = json.loads(resp.data.decode('utf8'))['packages']
-
-        assert data['own'] == []
-        assert data['shared'] == [dict(owner=self.user, name=self.pkg, is_public=False, is_team=True)]
-
-    @patch('quilt_server.views.ALLOW_ANONYMOUS_ACCESS', True)
-    @patch('quilt_server.views.ALLOW_TEAM_ACCESS', True)
-    def testTeamProfileWithPublic(self):
-        """
-        Test the profile endpoint but with teams *AND* public packages.
-        """
-        self.put_package(self.user, 'pkg1', RootNode(children=dict()), is_team=True)
-        self.put_package(self.user, 'pkg2', RootNode(children=dict()), is_public=True)
-        self.put_package(self.user, 'pkg3', RootNode(children=dict()), is_team=True, is_public=True)
-
-        # The user can see own packages.
-        resp = self.app.get(
-            '/api/profile',
-            headers={
-                'Authorization': self.user
-            }
-        )
-
-        assert resp.status_code == requests.codes.ok
-        data = json.loads(resp.data.decode('utf8'))['packages']
-
-        assert data['own'] == [
-            dict(owner=self.user, name='pkg1', is_public=False, is_team=True),
-            dict(owner=self.user, name='pkg2', is_public=True, is_team=False),
-            dict(owner=self.user, name='pkg3', is_public=True, is_team=True),
-            dict(owner=self.user, name=self.pkg, is_public=False, is_team=False),
-        ]
-        assert data['shared'] == []
-
-
-        # Other users can't see anything.
-        sharewith = "anotheruser"
-
-        resp = self.app.get(
-            '/api/profile',
-            headers={
-                'Authorization': sharewith
-            }
-        )
-
-        assert resp.status_code == requests.codes.ok
-        data = json.loads(resp.data.decode('utf8'))['packages']
-
-        assert data['own'] == []
-        assert data['shared'] == []
-
-        # Packages that are both team and shared show up under "shared".
-        resp = self._share_package(self.user, 'pkg1', sharewith)
-        assert resp.status_code == requests.codes.ok
-        resp = self._share_package(self.user, 'pkg3', sharewith)
-        assert resp.status_code == requests.codes.ok
-
-        resp = self.app.get(
-            '/api/profile',
-            headers={
-                'Authorization': sharewith
-            }
-        )
-
-        assert resp.status_code == requests.codes.ok
-        data = json.loads(resp.data.decode('utf8'))['packages']
-
-        assert data['own'] == []
-        assert data['shared'] == [
-            dict(owner=self.user, name='pkg1', is_public=False, is_team=True),
-            dict(owner=self.user, name='pkg3', is_public=True, is_team=True),
-        ]
-
-    @patch('quilt_server.views.ALLOW_ANONYMOUS_ACCESS', True)
     def testRecentPackages(self):
         # Push two public packages.
         for i in range(2):
@@ -737,149 +511,6 @@ class AccessTestCase(QuiltTestCase):
         names = [pkg['name'] for pkg in data['packages']]
         assert len(names) == 4
         assert set(names) == set(['pkg0', 'pkg1', 'pkg2', 'pkg3'])
-
-    @patch('quilt_server.views.ALLOW_ANONYMOUS_ACCESS', True)
-    def testBasicSearch(self):
-        for i in [1, 2]:
-            pkg = 'public%d' % i
-            self.put_package(self.user, pkg, RootNode(children=dict()), is_public=True, tag_latest=True)
-
-        self.put_package(self.user, 'private', RootNode(children=dict()), tag_latest=True)
-
-        def _test_query(query, headers, expected_results):
-            params = dict(q=query)
-            resp = self.app.get(
-                '/api/search/?%s' % urllib.parse.urlencode(params),
-                headers=headers
-            )
-
-            assert resp.status_code == requests.codes.ok
-            data = json.loads(resp.data.decode('utf8'))
-
-            results = ['%(owner)s/%(name)s' % pkg for pkg in data['packages']]
-            assert results == expected_results
-
-        _test_query("test_user/public1", {}, ["test_user/public1"])
-        _test_query("Test_User/Public1", {}, ["test_user/public1"])
-        _test_query("test_user/private", {}, [])
-        _test_query("test_user/", {}, ["test_user/public1", "test_user/public2"])
-        _test_query("test_user public1", {}, ["test_user/public1"])
-        _test_query("", {}, ["test_user/public1", "test_user/public2"])
-        _test_query("foo", {}, [])
-
-        auth = {'Authorization': self.user}
-
-        _test_query("test_user/public1", auth, ["test_user/public1"])
-        _test_query("private", auth, ["test_user/private"])
-        _test_query("test", auth, [
-            "test_user/private", "test_user/public1", "test_user/public2"
-        ])
-        _test_query("", auth, [
-            "test_user/private", "test_user/public1", "test_user/public2"
-        ])
-
-    @patch('quilt_server.views.ALLOW_ANONYMOUS_ACCESS', True)
-    def testBasicSearchOrder(self):
-        for pkg in ['a', 'B', 'c', 'D']:
-            self.put_package(self.user, pkg, RootNode(children=dict()), is_public=True, tag_latest=True)
-
-        params = dict(q=self.user)
-        resp = self.app.get(
-            '/api/search/?%s' % urllib.parse.urlencode(params)
-        )
-
-        assert resp.status_code == requests.codes.ok
-        data = json.loads(resp.data.decode('utf8'))
-
-        names = [pkg['name'] for pkg in data['packages']]
-        assert names == ['a', 'B', 'c', 'D']
-
-    @patch('quilt_server.views.ALLOW_ANONYMOUS_ACCESS', True)
-    def testSearchReadmeSnippet(self):
-        readme_contents = 'foo' * 1000
-        blob_hash = '8db466bdfc3265dd1347843b31ed34af0a0c2e6ff0fd4d6a5853755f0e68b8a0'
-
-        contents = RootNode(dict(
-            README=FileNode([blob_hash], dict())
-        ))
-
-        self._mock_object(self.user, blob_hash, readme_contents.encode())
-        self.put_package(self.user, 'pkg', contents, is_public=True, tag_latest=True)
-
-        resp = self.app.get('/api/search/?q=pkg')
-        assert resp.status_code == requests.codes.ok
-        data = json.loads(resp.data.decode('utf8'))
-        packages = data['packages']
-
-        assert len(packages) == 1
-        assert packages[0]['is_public'] is True
-        assert packages[0]['is_team'] is False
-        assert packages[0]['readme_preview'] == readme_contents[0:1024]
-
-    @patch('quilt_server.views.ALLOW_ANONYMOUS_ACCESS', True)
-    def testFullTextSearch(self):
-        # Packages with READMEs.
-        packages = {
-            "clinton_email": (
-                "On Monday, August 31, the State Department released nearly 7,000 pages of Clinton’s "
-                "heavily redacted emails (its biggest release of emails to date)."
-            ),
-            "wine": (
-                "From the UCI Machine Learning Repository: Lichman, M. (2013). UCI Machine Learning Repository. "
-                "Irvine, CA: University of California, School of Information and Computer Science."
-            ),
-            "dogscats": "",
-            "nothing": "There are no Clinton's emails or wine data here.",
-        }
-
-        for name, readme in packages.items():
-            blob_hash = hashlib.sha256(readme.encode()).hexdigest()
-            contents = RootNode(dict(
-                README=FileNode([blob_hash], dict())
-            ))
-
-            self._mock_object(self.user, blob_hash, readme.encode())
-            self.put_package(self.user, name, contents, is_public=True, tag_latest=True)
-
-        # Package with no README, but more nodes.
-        contents2 = RootNode(dict(
-            wine=GroupNode(dict()),
-            baz=GroupNode(dict())
-        ))
-        self.put_package(self.user, "foo", contents2, is_public=True, tag_latest=True)
-
-        def _test_query(query, headers, expected_results):
-            params = dict(q=query)
-            resp = self.app.get(
-                '/api/search/?%s' % urllib.parse.urlencode(params),
-                headers=headers
-            )
-
-            assert resp.status_code == requests.codes.ok
-            data = json.loads(resp.data.decode('utf8'))
-
-            results = ['%(owner)s/%(name)s' % pkg for pkg in data['packages']]
-            assert results == expected_results
-
-        # Stemming
-        _test_query("redact", {}, ["test_user/clinton_email"])
-        _test_query("releasing", {}, ["test_user/clinton_email"])
-
-        # Stemming on package name
-        _test_query("no wining", {}, ["test_user/wine", "test_user/foo", "test_user/nothing"])
-
-        # Multiple words
-        _test_query("state department's biggest release", {}, ["test_user/clinton_email"])
-
-        # Keywords in package name
-        _test_query("users dog cat", {}, ["test_user/dogscats"])
-
-        # Order precedence: package name, metadata, README
-        _test_query("clinton", {}, ["test_user/clinton_email", "test_user/nothing"])
-        _test_query("wine", {}, ["test_user/wine", "test_user/foo", "test_user/nothing"])
-
-        # Different keywords match different sources: package name and README.
-        _test_query("nothing wine", {}, ["test_user/nothing"])
 
     @patch('quilt_server.views.ALLOW_ANONYMOUS_ACCESS', True)
     def testGetObjects(self):
